backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pandas>=2.2.0
yfinance>=0.2.40
ta>=0.11.0
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.0
exchange-calendars>=4.5.0
